This is a random code.
Have you ever decided to go bowling with an indoor bowling set at your house but have been unable to agree on the rules? My Bowling Randomizer to the rescue! Just run it and you will have a distance to bowl from and a part that you can ignore: what ball you all will use. You could run this before every frame to decide the rules for that frame! Thanks for using it!